python_backend/requirements.txt
fastapi
uvicorn
railtracks
railtracks[cli]
ibm-watsonx-ai
ibm-db
python-dotenv
pydantic
numpy
requests
